@@ -1,6 +1,6 @@
 import { Router } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
-import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { UntypedFormGroup, UntypedFormControl, Validators } from '@angular/forms';
 import { Title } from '@angular/platform-browser';
 
 import { NotificationService } from 'src/app/core/services/notification.service';
@@ -14,7 +14,7 @@ import { AuthenticationService } from 'src/app/core/services/auth.service';
 export class PasswordResetRequestComponent implements OnInit {
 
   private email!: string;
-  form!: FormGroup;
+  form!: UntypedFormGroup;
   loading!: boolean;
 
   constructor(private authService: AuthenticationService,
@@ -25,8 +25,8 @@ export class PasswordResetRequestComponent implements OnInit {
   ngOnInit() {
     this.titleService.setTitle('angular-material-template - Password Reset Request');
 
-    this.form = new FormGroup({
-      email: new FormControl('', [Validators.required, Validators.email])
+    this.form = new UntypedFormGroup({
+      email: new UntypedFormControl('', [Validators.required, Validators.email])
     });
 
     this.form.get('email')?.valueChanges
