@@ -6,12 +6,11 @@ import { DashboardHomeComponent } from './dashboard-home/dashboard-home.componen
 import { SharedModule } from 'src/app/shared/shared.module';
 
 @NgModule({
-  declarations: [DashboardHomeComponent],
-  imports: [
-    CommonModule,
-    DashboardRoutingModule,
-    SharedModule
-  ],
-  entryComponents: []
+    declarations: [DashboardHomeComponent],
+    imports: [
+        CommonModule,
+        DashboardRoutingModule,
+        SharedModule
+    ]
 })
 export class DashboardModule { }
